설계 방식 변경. 1. 기존 es6전환 취소 1. 연결 방식 좀더 안정적으로 개선
