+++ app/custom_modules/router/scanner/serial.js
@@ -43,11 +43,11 @@ Scanner.prototype.scan = function(serialport, extension, config, callback) {
 		var duration = config.hardware.duration;
 		var firmwarecheck = config.hardware.firmwarecheck;
 		var pnpId = config.hardware.pnpId;
+		var type = config.hardware.type;
 		var checkComPort = config.select_com_port || false;
 		var myComPort = config.this_com_port;
-		var type = config.hardware.type;
 
-		if(checkComPort && !myComPort)  {
+		if((checkComPort || type === 'bluetooth') && !myComPort)  {
 			self.router.emit('state', 'select_port', devices);
 			callback('select_port');
 			return;
@@ -106,15 +106,15 @@ Scanner.prototype.scan = function(serialport, extension, config, callback) {
 				var connector = self.connectors[comName];
 				if(connector == undefined) {
 					connector = require('../connector/serial').create();
+					self.connectors[comName] = connector;
 					connector.open(comName, config.hardware, function(error, sp) {
 						if(error) {
+							delete self.connectors[comName];
 							if(callback) {
 								callback(error);
 							}
 						} else {
-							// sp.flush();
 							self.setConnector(connector);
-							self.connectors[comName] = connector;
 							if(control) {
 								var flashFirmware;
 								if(firmwarecheck) {
