+++ app/package.json
@@ -2,8 +2,6 @@
   "name": "Entry_HW",
   "version": "1.5.11",
   "description": "entry hardware connect program",
-  "main": "es6-init.js",
-  "updateServer": "play02.play-entry.com:4001",
   "dependencies": {
     "es6-promise": "^3.2.1",
     "object.assign": "^4.0.4",
