--- src/es6-init.js
@@ -1,3 +1,0 @@
-var path = require('path');
-console.log(__dirname);
-require('electron-compile').init(__dirname, './main', true);
