--- src/main.js
@@ -1,4 +1,0 @@
-'use strict';
-
-import electron from 'electron';
-console.log(electron);
